Move class to better package
